Strip all formatting logic from LogAggregator into Formatter interface
