+++ pkg/skaffold/kubernetes/logger/log_test.go
@@ -25,82 +25,9 @@ import (
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
-	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/graph"
 	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/kubernetes"
-	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/output"
-	latestV1 "github.com/GoogleContainerTools/skaffold/pkg/skaffold/schema/latest/v1"
 	"github.com/GoogleContainerTools/skaffold/testutil"
 )
-
-func TestColorForPod(t *testing.T) {
-	tests := []struct {
-		description   string
-		pod           *v1.Pod
-		expectedColor output.Color
-	}{
-		{
-			description:   "not found",
-			pod:           &v1.Pod{},
-			expectedColor: output.None,
-		},
-		{
-			description: "found",
-			pod: &v1.Pod{
-				Spec: v1.PodSpec{
-					Containers: []v1.Container{
-						{Image: "image"},
-					},
-				},
-			},
-			expectedColor: output.DefaultColorCodes[0],
-		},
-		{
-			description: "ignore tag",
-			pod: &v1.Pod{
-				Spec: v1.PodSpec{
-					Containers: []v1.Container{
-						{Image: "image:tag"},
-					},
-				},
-			},
-			expectedColor: output.DefaultColorCodes[0],
-		},
-		{
-			description: "second image",
-			pod: &v1.Pod{
-				Spec: v1.PodSpec{
-					Containers: []v1.Container{
-						{Image: "second:tag"},
-					},
-				},
-			},
-			expectedColor: output.DefaultColorCodes[1],
-		},
-		{
-			description: "accept image with digest",
-			pod: &v1.Pod{
-				Spec: v1.PodSpec{
-					Containers: []v1.Container{
-						{Image: "second:tag@sha256:d3f4dd1541ee34b96850efc46955bada1a415b0594dc9948607c0197d2d16749"},
-					},
-				},
-			},
-			expectedColor: output.DefaultColorCodes[1],
-		},
-	}
-
-	l := NewLogAggregator(nil, nil, nil, nil)
-	// artifacts are registered using their tag, since these have default repo substitutions applied
-	l.RegisterArtifacts([]graph.Artifact{{Tag: "image"}, {Tag: "second"}})
-
-	for _, test := range tests {
-		testutil.Run(t, test.description, func(t *testutil.T) {
-			color := l.PodColor(test.pod)
-
-			t.CheckTrue(test.expectedColor == color)
-		})
-	}
-}
 
 func TestSinceSeconds(t *testing.T) {
 	tests := []struct {
@@ -176,70 +103,6 @@ func TestLogAggregatorZeroValue(t *testing.T) {
 	m.Stop()
 }
 
-func TestPrefix(t *testing.T) {
-	tests := []struct {
-		description    string
-		prefix         string
-		pod            v1.Pod
-		container      v1.ContainerStatus
-		expectedPrefix string
-	}{
-		{
-			description:    "auto (different names)",
-			prefix:         "auto",
-			pod:            podWithName("pod"),
-			container:      containerWithName("container"),
-			expectedPrefix: "[pod container]",
-		},
-		{
-			description:    "auto (same names)",
-			prefix:         "auto",
-			pod:            podWithName("hello"),
-			container:      containerWithName("hello"),
-			expectedPrefix: "[hello]",
-		},
-		{
-			description:    "container",
-			prefix:         "container",
-			pod:            podWithName("pod"),
-			container:      containerWithName("container"),
-			expectedPrefix: "[container]",
-		},
-		{
-			description:    "podAndContainer (different names)",
-			prefix:         "podAndContainer",
-			pod:            podWithName("pod"),
-			container:      containerWithName("container"),
-			expectedPrefix: "[pod container]",
-		},
-		{
-			description:    "podAndContainer (same names)",
-			prefix:         "podAndContainer",
-			pod:            podWithName("hello"),
-			container:      containerWithName("hello"),
-			expectedPrefix: "[hello hello]",
-		},
-		{
-			description:    "none",
-			prefix:         "none",
-			pod:            podWithName("hello"),
-			container:      containerWithName("hello"),
-			expectedPrefix: "",
-		},
-	}
-	for _, test := range tests {
-		testutil.Run(t, test.description, func(t *testutil.T) {
-			logger := NewLogAggregator(nil, nil, nil, &mockConfig{log: latestV1.LogsConfig{
-				Prefix: test.prefix,
-			}})
-
-			p := logger.prefix(&test.pod, test.container)
-
-			t.CheckDeepEqual(test.expectedPrefix, p)
-		})
-	}
-}
-
 func podWithName(n string) v1.Pod {
 	return v1.Pod{
 		ObjectMeta: metav1.ObjectMeta{
@@ -252,24 +115,4 @@ func containerWithName(n string) v1.ContainerStatus {
 	return v1.ContainerStatus{
 		Name: n,
 	}
-}
-
-type mockConfig struct {
-	log latestV1.LogsConfig
-}
-
-func (c *mockConfig) Tail() bool {
-	return true
-}
-
-func (c *mockConfig) PipelineForImage(string) (latestV1.Pipeline, bool) {
-	var pipeline latestV1.Pipeline
-	pipeline.Deploy.Logs = c.log
-	return pipeline, true
-}
-
-func (c *mockConfig) DefaultPipeline() latestV1.Pipeline {
-	var pipeline latestV1.Pipeline
-	pipeline.Deploy.Logs = c.log
-	return pipeline
 }
