+++ pkg/skaffold/log/stream/stream.go
@@ -21,23 +21,19 @@ import (
 	"context"
 	"fmt"
 	"io"
-	"sync"
 
 	"github.com/sirupsen/logrus"
 
-	eventV2 "github.com/GoogleContainerTools/skaffold/pkg/skaffold/event/v2"
-	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/output"
+	"github.com/GoogleContainerTools/skaffold/pkg/skaffold/log"
 )
 
 //nolint:golint
-func StreamRequest(ctx context.Context, out io.Writer, headerColor output.Color, prefix, podName, containerName string, stopper chan bool, lock sync.Locker, isMuted func() bool, rc io.Reader) error {
+func StreamRequest(ctx context.Context, out io.Writer, formatter log.Formatter, rc io.Reader) error {
 	r := bufio.NewReader(rc)
 	for {
 		select {
 		case <-ctx.Done():
-			logrus.Infof("%s interrupted", prefix)
-			return nil
-		case <-stopper:
+			logrus.Infof("%s interrupted", formatter.Name())
 			return nil
 		default:
 			// Read up to newline
@@ -49,25 +45,7 @@ func StreamRequest(ctx context.Context, out io.Writer, headerColor output.Color,
 				return fmt.Errorf("reading bytes from log stream: %w", err)
 			}
 
-			printLogLine(headerColor, out, isMuted, lock, podName, containerName, prefix, line)
+			formatter.PrintLine(out, line)
 		}
-	}
-}
-
-func printLogLine(headerColor output.Color, out io.Writer, isMuted func() bool, lock sync.Locker, podName, containerName, prefix, text string) {
-	formattedPrefix := prefix
-	if output.IsColorable(out) {
-		formattedPrefix = headerColor.Sprintf("%s", prefix)
-	}
-	formattedLine := fmt.Sprintf("%s %s", formattedPrefix, text)
-	eventV2.ApplicationLog(podName, containerName, formattedPrefix, text, formattedLine)
-
-	if !isMuted() {
-		lock.Lock()
-
-		headerColor.Fprintf(out, "%s ", prefix)
-		fmt.Fprint(out, text)
-
-		lock.Unlock()
 	}
 }
